[luci] Extract visitors to quantize activation
This extracts visitors to quantize activation from QuantizeWithMinMax.cpp.
- QuantizeActivation
- QuantizeConstInputActivation
- QuantizeSpecialActivation

ONE-DCO-1.0-Signed-off-by: Hyukjin Jeong <hj1.jeong@samsung.com>
