+++ compiler/luci/pass/src/QuantizeWithMinMaxPass.cpp
@@ -17,6 +17,7 @@
 #include "luci/Pass/QuantizeWithMinMaxPass.h"
 #include "luci/Pass/PropagateQParamForwardPass.h"
 #include "luci/Pass/PropagateQParamBackwardPass.h"
+#include "QuantizeActivation.h"
 #include "QuantizeWeights.h"
 #include "QuantizationUtils.h"
 #include "ProgressReporter.h"
@@ -153,56 +154,6 @@ void quant_const(CircleConst *node, loco::DataType quant_type)
   node->quantparam(std::move(quantparam));
 }
 
-// Check if the node is the bias of Conv2D, DepthwiseConv2D, FullyConnected, or TransposeConv layer
-// Returns a list of <input, weights, output> vectors for the above operators.
-// Note that it returns a 'list' because bias can be used by multiple operators.
-std::vector<std::vector<loco::Node *>> get_input_weight_output_of_bias(CircleNode *node)
-{
-  std::vector<std::vector<loco::Node *>> result;
-  auto circle_const = dynamic_cast<CircleConst *>(node);
-  if (circle_const == nullptr)
-    return result;
-
-  auto succs = loco::succs(node);
-
-  for (auto out : succs)
-  {
-    auto conv = dynamic_cast<CircleConv2D *>(out);
-    if (conv != nullptr && conv->bias() == circle_const)
-    {
-      assert(conv->input() != nullptr);
-      assert(conv->filter() != nullptr);
-      result.push_back({conv->input(), conv->filter(), conv});
-      continue;
-    }
-    auto dw_conv = dynamic_cast<CircleDepthwiseConv2D *>(out);
-    if (dw_conv != nullptr && dw_conv->bias() == circle_const)
-    {
-      assert(dw_conv->input() != nullptr);
-      assert(dw_conv->filter() != nullptr);
-      result.push_back({dw_conv->input(), dw_conv->filter(), dw_conv});
-      continue;
-    }
-    auto fc = dynamic_cast<CircleFullyConnected *>(out);
-    if (fc != nullptr && fc->bias() == circle_const)
-    {
-      assert(fc->input() != nullptr);
-      assert(fc->weights() != nullptr);
-      result.push_back({fc->input(), fc->weights(), fc});
-      continue;
-    }
-    auto tconv = dynamic_cast<CircleTransposeConv *>(out);
-    if (tconv != nullptr && tconv->bias() == circle_const)
-    {
-      assert(tconv->outBackprop() != nullptr);
-      assert(tconv->filter() != nullptr);
-      result.push_back({tconv->outBackprop(), tconv->filter(), tconv});
-      continue;
-    }
-  }
-  return result;
-}
-
 CircleConst *asym_quant_bias_per_layer(CircleConst *node, float input_scale, float weight_scale,
                                        float *scaling_factor, int64_t *zp)
 {
@@ -291,183 +242,6 @@ CircleConst *int16_quant_bias_per_channel(CircleConst *node, float input_scale,
 
   return new_bias;
 }
-
-bool has_min_max(const CircleNode *node)
-{
-  return node->quantparam() && !node->quantparam()->min.empty() && !node->quantparam()->max.empty();
-}
-
-void set_act_qparam(luci::CircleNode *node, float scale, int64_t zp)
-{
-  assert(node);               // FIX_CALLER_UNLESS
-  assert(node->quantparam()); // FIX_CALLER_UNLESS
-
-  auto qparam = node->quantparam();
-  assert(qparam->scale.size() == 1); // FIX_CALLER_UNLESS
-  assert(qparam->zerop.size() == 1); // FIX_CALLER_UNLESS
-  qparam->scale[0] = scale;
-  qparam->zerop[0] = zp;
-}
-
-// For nodes with integer output, we use integer scale
-void set_int_scale(luci::CircleNode *node)
-{
-  assert(node); // FIX_CALLER_UNLESS
-
-  auto qparam = node->quantparam();
-  assert(qparam);                    // FIX_CALLER_UNLESS
-  assert(qparam->scale.size() == 1); // FIX_CALLER_UNLESS
-
-  auto fp_scale = qparam->scale[0];
-  qparam->scale[0] = fp_scale < 1 ? 1.0f : std::round(fp_scale);
-}
-
-/**
- * @brief Manually set scale/zp of output tensor of special Ops
- */
-struct QuantizeSpecialActivation final : public luci::CircleNodeMutableVisitor<void>
-{
-  QuantizeSpecialActivation(loco::DataType input, loco::DataType output)
-    : input_type(input), output_type(output)
-  {
-  }
-
-  loco::DataType input_type;
-  loco::DataType output_type;
-
-  void visit(luci::CircleNode *node)
-  {
-    // Nodes fused with activation functions which need special quantization
-    auto fused_act_node = dynamic_cast<CircleNodeMixin<CircleNodeTrait::FusedActFunc> *>(node);
-    if (fused_act_node != nullptr &&
-        fused_act_node->fusedActivationFunction() == FusedActFunc::TANH)
-    {
-      if (output_type == loco::DataType::U8)
-        set_act_qparam(node, 2.0f / 256.0f, 128);
-      else
-      {
-        assert(output_type == loco::DataType::S16);
-        set_act_qparam(node, 1.0f / 32768.0f, 0);
-      }
-    }
-  }
-
-  void visit(luci::CircleLogistic *node)
-  {
-    if (output_type == loco::DataType::U8)
-      set_act_qparam(node, 1.0f / 256.0f, 0);
-    else
-    {
-      assert(output_type == loco::DataType::S16);
-      set_act_qparam(node, 1.0f / 32768.0f, 0);
-    }
-  }
-
-  void visit(luci::CircleTanh *node)
-  {
-    if (output_type == loco::DataType::U8)
-      set_act_qparam(node, 2.0f / 256.0f, 128);
-    else
-    {
-      assert(output_type == loco::DataType::S16);
-      set_act_qparam(node, 1.0f / 32768.0f, 0);
-    }
-  }
-
-  void visit(luci::CircleSoftmax *node)
-  {
-    if (output_type == loco::DataType::U8)
-      set_act_qparam(node, 1.0f / 255.0f, 0);
-    else
-    {
-      assert(output_type == loco::DataType::S16);
-      set_act_qparam(node, 1.0f / 32767.0f, 0);
-    }
-  }
-
-  void visit(luci::CircleFloor *node) { set_int_scale(node); }
-  void visit(luci::CircleFloorDiv *node) { set_int_scale(node); }
-  void visit(luci::CircleFloorMod *node) { set_int_scale(node); }
-  void visit(luci::CircleCeil *node) { set_int_scale(node); }
-};
-
-/**
- * @brief QuantizeActivation quantizes tensors for activations
- * @details Quantize using recorded min/max values
- */
-struct QuantizeActivation final : public luci::CircleNodeMutableVisitor<void>
-{
-  QuantizeActivation(loco::DataType input, loco::DataType output)
-    : input_type(input), output_type(output)
-  {
-  }
-
-  loco::DataType input_type;
-  loco::DataType output_type;
-
-  // Quantize input tensors of each node
-  void visit(luci::CircleNode *node)
-  {
-    LOGGER(l);
-    INFO(l) << "QuantizeActivation visit node: " << node->name() << std::endl;
-
-    // Check if this is already quantized
-    if (is_quantized(node))
-      return;
-
-    // Check if this is bias (bias is quantized later)
-    auto iwo = get_input_weight_output_of_bias(node);
-    if (iwo.size() > 0)
-      return;
-
-    // Check if this is bool type (bool type is not quantized)
-    if (node->dtype() == loco::DataType::BOOL)
-      return;
-
-    // Check if this is activation
-    // We assume min/max are recorded only for activations
-    if (has_min_max(node) && !is_weights(node))
-    {
-      // Quantize using recorded min/max
-      auto quantparam = node->quantparam();
-      assert(quantparam);
-      assert(quantparam->min.size() == 1); // only support layer-wise quant
-      assert(quantparam->max.size() == 1); // only support layer-wise quant
-      auto min = quantparam->min[0];
-      auto max = quantparam->max[0];
-
-      float scaling_factor{0};
-      int64_t zp{0};
-      float nudged_min{0};
-      float nudged_max{0};
-
-      if (output_type == loco::DataType::U8)
-      {
-        compute_asym_scale_zp(min, max, scaling_factor, zp, nudged_min, nudged_max);
-        node->dtype(loco::DataType::U8);
-      }
-      else
-      {
-        compute_sym_scale_zp(min, max, scaling_factor, zp, nudged_min, nudged_max);
-        node->dtype(loco::DataType::S16);
-      }
-
-      node->quantparam()->scale.push_back(scaling_factor);
-      node->quantparam()->zerop.push_back(zp);
-    }
-    // Fix special attributes
-    if (node->opcode() == luci::CircleOpcode::CAST)
-    {
-      auto *cast = loco::must_cast<luci::CircleCast *>(node);
-      auto *cast_input = loco::must_cast<luci::CircleNode *>(cast->x());
-
-      // make sure that cast_input is already quantized
-      assert(cast_input->dtype() != loco::DataType::FLOAT32);
-      cast->in_data_type(cast_input->dtype());
-      cast->out_data_type(cast->dtype());
-    }
-  }
-};
 
 struct QuantizeBias final : public luci::CircleNodeMutableVisitor<void>
 {
@@ -633,165 +407,6 @@ private:
   }
 
   void visit(luci::CircleNode *) {}
-};
-
-// Quantize constant input activation of a node
-// The input of a node is quantized if it is
-// 1. Constant (instance of CircleConst*)
-// 2. Activation (other inputs e.g., weights, bias, axis, etc should not be quantized here)
-struct QuantizeConstInputActivation final : public luci::CircleNodeMutableVisitor<void>
-{
-  QuantizeConstInputActivation(loco::DataType output_type) : _output_type(output_type) {}
-
-private:
-  loco::DataType _output_type;
-
-// Skip NODE
-#define SKIP(NODE) \
-  void visit(NODE *) {}
-
-// INPUT_NAME is the only activation of NODE
-#define QUANTIZE_SINGLE_CONST_INPUT(NODE, INPUT_NAME)           \
-  void visit(NODE *node)                                        \
-  {                                                             \
-    auto input = node->INPUT_NAME();                            \
-    auto const_node = dynamic_cast<luci::CircleConst *>(input); \
-    if (const_node && !is_quantized(const_node))                \
-    {                                                           \
-      auto new_const = luci::clone(const_node);                 \
-      quant_const(new_const, _output_type);                     \
-      node->INPUT_NAME(new_const);                              \
-    }                                                           \
-  }
-
-// INPUT_NAME1 and INPUT_NAME2 are the only activations of NODE
-#define QUANTIZE_TWO_CONST_INPUTS(NODE, INPUT_NAME1, INPUT_NAME2) \
-  void visit(NODE *node)                                          \
-  {                                                               \
-    auto input1 = node->INPUT_NAME1();                            \
-    auto const_node1 = dynamic_cast<luci::CircleConst *>(input1); \
-    if (const_node1 && !is_quantized(const_node1))                \
-    {                                                             \
-      auto new_const1 = luci::clone(const_node1);                 \
-      quant_const(new_const1, _output_type);                      \
-      node->INPUT_NAME1(new_const1);                              \
-    }                                                             \
-    auto input2 = node->INPUT_NAME2();                            \
-    auto const_node2 = dynamic_cast<luci::CircleConst *>(input2); \
-    if (const_node2 && !is_quantized(const_node2))                \
-    {                                                             \
-      auto new_const2 = luci::clone(const_node2);                 \
-      quant_const(new_const2, _output_type);                      \
-      node->INPUT_NAME2(new_const2);                              \
-    }                                                             \
-  }
-
-  // Default behavior (NYI)
-  void visit(luci::CircleNode *node)
-  {
-    for (uint32_t i = 0; i < node->arity(); i++)
-    {
-      auto input_node = node->arg(i);
-      auto const_node = dynamic_cast<luci::CircleConst *>(input_node);
-      if (const_node != nullptr)
-        throw std::runtime_error("Unsupported Op for const inputs");
-    }
-  }
-
-  // Handled in QuantizeWeights and QuantizeBias
-  SKIP(luci::CircleConv2D)
-  SKIP(luci::CircleDepthwiseConv2D)
-  SKIP(luci::CircleFullyConnected)
-  SKIP(luci::CircleInstanceNorm)
-  SKIP(luci::CirclePRelu)
-  SKIP(luci::CircleTransposeConv)
-
-  // Handled in PropagateQParamBackwardPass
-  SKIP(luci::CircleConcatenation)
-  SKIP(luci::CirclePadV2)
-  SKIP(luci::CirclePack)
-  SKIP(luci::CircleOneHot)
-
-  // Inputs of logical Ops are bool, thus not quantized
-  SKIP(luci::CircleLogicalOr)
-  SKIP(luci::CircleLogicalAnd)
-  SKIP(luci::CircleLogicalNot)
-
-  // Ops that receive a single activation as an input
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleArgMax, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleArgMin, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleBatchToSpaceND, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleDepthToSpace, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleElu, features)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleExp, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleFloor, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleGather, params)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleLocalResponseNormalization, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleLogistic, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleMean, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleMirrorPad, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CirclePad, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReduceAny, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReduceProd, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReduceMax, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReduceMin, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReshape, tensor)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleResizeBilinear, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleResizeNearestNeighbor, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleReverseSequence, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleRsqrt, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSlice, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSoftmax, logits)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSpaceToBatchND, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSpaceToDepth, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSplit, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSplitV, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSqrt, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleStridedSlice, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleSum, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleTanh, x)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleTile, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleTopKV2, input)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleTranspose, a)
-  QUANTIZE_SINGLE_CONST_INPUT(luci::CircleUnpack, value)
-
-  // Ops that receive two activations as inputs
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleAdd, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleBatchMatMul, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleDiv, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleEqual, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleFloorDiv, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleGreater, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleGreaterEqual, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleLess, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleLessEqual, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleMaximum, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleMinimum, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleMul, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleNotEqual, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CirclePow, x, y)
-  QUANTIZE_TWO_CONST_INPUTS(luci::CircleSub, x, y)
-
-  // AddN has arbitrary number of inputs
-  void visit(luci::CircleAddN *node)
-  {
-    auto arity = node->arity();
-    for (uint32_t i = 0; i < arity; i++)
-    {
-      auto input_node = node->inputs(i);
-      auto const_node = dynamic_cast<luci::CircleConst *>(input_node);
-      if (const_node && !is_quantized(const_node))
-      {
-        auto new_const = luci::clone(const_node);
-        quant_const(new_const, _output_type);
-        node->inputs(i, new_const);
-      }
-    }
-  }
-
-#undef SKIP
-#undef QUANTIZE_SINGLE_CONST_INPUT
-#undef QUANTIZE_TWO_CONST_INPUTS
 };
 
 } // namespace
